requirements.txt
# Use Python 3.10–3.13; CrewAI does not support Python 3.14 yet.
crewai>=0.86.0,<1.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
PyYAML>=6.0.1
pytest>=8.0.0
matplotlib>=3.8.0
